Ciosmaps update
The ciosmaps.xml file now includes the base IOS revision number
in the data section (u16 at 0x30) so that tools like Syscheck can
display it for verification.

Also, there are new "console" and "region" flags so that certain
cIOS entries will only be displayed on certain consoles. Support
for that still needs to be added to the d2x cIOS installer.